+++ lib/dnsimple/struct.rb
@@ -20,7 +20,7 @@ require_relative 'struct/domain_check'
 require_relative 'struct/email_forward'
 require_relative 'struct/extended_attribute'
 require_relative 'struct/oauth_token'
-require_relative 'struct/record'
+require_relative 'struct/zone_record'
 require_relative 'struct/service'
 require_relative 'struct/template'
 require_relative 'struct/template_record'
