+++ lib/dnsimple/struct/zone_record.rb
